+++ tests/conftest.py
@@ -1,4 +1,7 @@
+import factory
+import pytest
 from pytest_factoryboy import register
+from rest_framework.test import APIClient
 
 from django.conf import settings
 
@@ -17,3 +20,34 @@ def pytest_configure():
     """
     settings.CELERY_TASK_ALWAYS_EAGER = True
     settings.EMAIL_BACKEND = "django.core.mail.backends.console.EmailBackend"
+
+
+@pytest.fixture
+def api_client():
+    return APIClient()
+
+
+@pytest.fixture
+def api_client_authenticate(db, user_factory, api_client):
+    user = user_factory()
+    api_client.force_authenticate(user=user)
+    yield api_client
+    api_client.force_authenticate(user=None)
+
+
+@pytest.fixture
+def create_job(db, job_factory):
+    return job_factory
+
+
+@pytest.fixture
+def create_job_as_dict(db, job_factory) -> dict:
+    return factory.build(dict, FACTORY_CLASS=JobFactory)
+
+
+@pytest.fixture
+def create_jobs(db, job_factory):
+    def _create_jobs(size=10):
+        return job_factory.simple_generate_batch(create=True, size=size)
+
+    return _create_jobs
