+++ tests/factories.py
@@ -6,7 +6,7 @@ from factory import fuzzy
 from faker import Faker
 
 from accounts.models import User
-from jobsapp.models import Job
+from jobsapp.models import NO_REMOTE, PARTIAL_REMOTE, REMOTE, Job
 
 faker = Faker("es_ES")
 
@@ -28,10 +28,11 @@ class JobFactory(factory.django.DjangoModelFactory):  # type: ignore
     user = factory.SubFactory("tests.factories.UserFactory")
     title = factory.Sequence(lambda n: f"Title {n}")
     location = faker.address()
-    company_name = faker.address()
+    company_name = faker.company()
     company_description = faker.text()
     description = factory.Sequence(lambda n: f"Description {n}")
     last_date = datetime.datetime.now() + datetime.timedelta(days=10)
     website = faker.url()
     type = "1"
     category = fuzzy.FuzzyChoice(["Senior", "Junior", "Manager"])
+    remote = fuzzy.FuzzyChoice([REMOTE, NO_REMOTE, PARTIAL_REMOTE])
